// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/tdee-insight-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface TdeeInsightHttpResponse {
    adaptiveTdee?: number | null;
    bmr?: number | null;
    confidence?: string;
    currentCalorieTarget?: number | null;
    dataDaysUsed?: number;
    estimatedTdee?: number | null;
    goalAdjustmentHint?: string | null;
    suggestedCalorieTarget?: number | null;
    weightTrendPerWeek?: number | null;
}